[I25-291] refactor: 폴더 구조 변경
