+++ src/app/(box-layout)/portfolio/page.tsx
@@ -1,4 +1,4 @@
-import SearchTab from './SearchTab';
+import SearchTab from '@/app/components/portfolio/SearchTab';
 import { getPersonalPortfolioData } from '@/services/portfolio/getPersonalPortfolioData.server';
 import getJobs from '@/services/portfolio/getJobs.server';
