+++ src/services/portfolio/getPaginatedPortfolioData.server.ts
@@ -2,7 +2,7 @@ import { createClient } from '@/services/supabase/client';
 import {
   PortfolioData,
   ProfileWithProjects,
-} from '@/app/(box-layout)/portfolio/types';
+} from '@/app/components/portfolio/types';
 import { PaginatedPortfolioResponse } from '@/types/pagination';
 
 export async function getPaginatedPortfolioData(
