+++ src/types/pagination.ts
@@ -1,4 +1,4 @@
-import { PortfolioData } from '@/app/(box-layout)/portfolio/types';
+import { PortfolioData } from '@/app/components/portfolio/types';
 
 export interface PaginatedResponse<T> {
   data: T[];
